+++ src/test/java/blackjack/domain/betting/BettingMoneyTest.java
@@ -5,6 +5,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 public class BettingMoneyTest {
 
     @DisplayName("게임 결과가 블랙잭이면 배팅 금액의 1.5배를 수익으로 받는다.")
@@ -14,10 +16,10 @@ public class BettingMoneyTest {
         BettingMoney bettingMoney = new BettingMoney(10000);
 
         // when
-        int profit = bettingMoney.calculateProfit(GameResult.BLACKJACK);
+        BigDecimal profit = bettingMoney.calculateProfit(GameResult.BLACKJACK);
 
         // then
-        Assertions.assertThat(profit).isEqualTo(15000);
+        Assertions.assertThat(profit).isEqualTo(BigDecimal.valueOf(15000.0));
     }
 
     @DisplayName("게임 결과가 블랙잭이면 배팅 금액의 0배를 수익으로 받는다.")
@@ -27,10 +29,10 @@ public class BettingMoneyTest {
         BettingMoney bettingMoney = new BettingMoney(10000);
 
         // when
-        int profit = bettingMoney.calculateProfit(GameResult.DRAW);
+        BigDecimal profit = bettingMoney.calculateProfit(GameResult.DRAW);
 
         // then
-        Assertions.assertThat(profit).isEqualTo(0);
+        Assertions.assertThat(profit).isEqualTo(BigDecimal.valueOf(0));
     }
 
     @DisplayName("게임 결과가 블랙잭이면 배팅 금액의 1배를 수익으로 받는다.")
@@ -40,11 +42,11 @@ public class BettingMoneyTest {
         BettingMoney bettingMoney = new BettingMoney(10000);
 
         // when
-        int profit = bettingMoney.calculateProfit(GameResult.WIN);
+        BigDecimal profit = bettingMoney.calculateProfit(GameResult.WIN);
 
 
         // then
-        Assertions.assertThat(profit).isEqualTo(10000);
+        Assertions.assertThat(profit).isEqualTo(BigDecimal.valueOf(10000));
     }
 
     @DisplayName("게임 결과가 블랙잭이면 배팅 금액의 -1배를 수익으로 받는다.")
@@ -54,9 +56,9 @@ public class BettingMoneyTest {
         BettingMoney bettingMoney = new BettingMoney(10000);
 
         // when
-        int profit = bettingMoney.calculateProfit(GameResult.LOSE);
+        BigDecimal profit = bettingMoney.calculateProfit(GameResult.LOSE);
 
         // then
-        Assertions.assertThat(profit).isEqualTo(-10000);
+        Assertions.assertThat(profit).isEqualTo(BigDecimal.valueOf(-10000));
     }
 }
